+++ HPXMLtoOpenStudio/tests/test_schedules.rb
@@ -187,6 +187,38 @@ class HPXMLtoOpenStudioSchedulesTest < Minitest::Test
     assert_in_epsilon(8760 * get_available_hrs_ratio(unavailable_month_hrs), get_annual_equivalent_full_load_hrs(model, Constants::ObjectTypeMechanicalVentilationHouseFan + ' schedule'), @tol)
   end
 
+  def test_simple_power_outage_schedules_overnight
+    args_hash = {}
+    hpxml_path = File.absolute_path(File.join(sample_files_dir, 'base-schedules-simple-power-outage.xml'))
+    hpxml = HPXML.new(hpxml_path: hpxml_path)
+    hpxml.header.unavailable_periods[0].begin_month = 12
+    hpxml.header.unavailable_periods[0].begin_day = 14
+    hpxml.header.unavailable_periods[0].begin_hour = 20
+    hpxml.header.unavailable_periods[0].end_month = 12
+    hpxml.header.unavailable_periods[0].end_day = 15
+    hpxml.header.unavailable_periods[0].end_hour = 6
+    XMLHelper.write_file(hpxml.to_doc(), @tmp_hpxml_path)
+    args_hash['hpxml_path'] = @tmp_hpxml_path
+    model, _hpxml, _hpxml_bldg = _test_measure(args_hash)
+
+    unavailable_month_hrs = { 11 => 10.0 }
+
+    assert_in_epsilon(6020, get_annual_equivalent_full_load_hrs(model, Constants::ObjectTypeOccupants + ' schedule'), @tol)
+    assert_in_epsilon(3049 * get_available_hrs_ratio(unavailable_month_hrs, @default_schedules_csv_data[SchedulesFile::Columns[:LightingInterior].name]['InteriorMonthlyScheduleMultipliers']), get_annual_equivalent_full_load_hrs(model, Constants::ObjectTypeLightingInterior + ' schedule'), @tol)
+    assert_in_epsilon(2895 * get_available_hrs_ratio(unavailable_month_hrs, @default_schedules_csv_data[SchedulesFile::Columns[:LightingInterior].name]['InteriorMonthlyScheduleMultipliers']), get_annual_equivalent_full_load_hrs(model, Constants::ObjectTypeLightingExterior + ' schedule'), @tol)
+    assert_in_epsilon(6673 * get_available_hrs_ratio(unavailable_month_hrs, @default_schedules_csv_data[SchedulesFile::Columns[:Refrigerator].name]['MonthlyScheduleMultipliers']), get_annual_equivalent_full_load_hrs(model, Constants::ObjectTypeRefrigerator + ' schedule'), @tol)
+    assert_in_epsilon(2441 * get_available_hrs_ratio(unavailable_month_hrs, @default_schedules_csv_data[SchedulesFile::Columns[:CookingRange].name]['MonthlyScheduleMultipliers']), get_annual_equivalent_full_load_hrs(model, Constants::ObjectTypeCookingRange + ' schedule'), @tol)
+    assert_in_epsilon(3285 * get_available_hrs_ratio(unavailable_month_hrs, @default_schedules_csv_data[SchedulesFile::Columns[:Dishwasher].name]['MonthlyScheduleMultipliers']), get_annual_equivalent_full_load_hrs(model, Constants::ObjectTypeDishwasher + ' schedule'), @tol)
+    assert_in_epsilon(4248 * get_available_hrs_ratio(unavailable_month_hrs, @default_schedules_csv_data[SchedulesFile::Columns[:ClothesWasher].name]['MonthlyScheduleMultipliers']), get_annual_equivalent_full_load_hrs(model, Constants::ObjectTypeClothesWasher + ' schedule'), @tol)
+    assert_in_epsilon(4502 * get_available_hrs_ratio(unavailable_month_hrs, @default_schedules_csv_data[SchedulesFile::Columns[:ClothesDryer].name]['MonthlyScheduleMultipliers']), get_annual_equivalent_full_load_hrs(model, Constants::ObjectTypeClothesDryer + ' schedule'), @tol)
+    assert_in_epsilon(6880 * get_available_hrs_ratio(unavailable_month_hrs, @default_schedules_csv_data[SchedulesFile::Columns[:PlugLoadsOther].name]['MonthlyScheduleMultipliers']), get_annual_equivalent_full_load_hrs(model, Constants::ObjectTypeMiscPlugLoads + ' schedule'), @tol)
+    assert_in_epsilon(3373 * get_available_hrs_ratio(unavailable_month_hrs, @default_schedules_csv_data[SchedulesFile::Columns[:PlugLoadsTV].name]['MonthlyScheduleMultipliers']), get_annual_equivalent_full_load_hrs(model, Constants::ObjectTypeMiscTelevision + ' schedule'), @tol)
+    assert_in_epsilon(4204 * get_available_hrs_ratio(unavailable_month_hrs, @default_schedules_csv_data[SchedulesFile::Columns[:HotWaterFixtures].name]['WaterFixturesMonthlyScheduleMultipliers']), get_annual_equivalent_full_load_hrs(model, Constants::ObjectTypeFixtures + ' schedule'), @tol)
+    assert_in_epsilon(4244 * get_available_hrs_ratio(unavailable_month_hrs, @default_schedules_csv_data[SchedulesFile::Columns[:HotWaterRecirculationPump].name]['RecirculationPumpMonthlyScheduleMultipliers']), get_annual_equivalent_full_load_hrs(model, Constants::ObjectTypeHotWaterRecircPump + ' schedule'), @tol)
+    assert_in_epsilon(5000, get_annual_equivalent_full_load_hrs(model, Constants::ObjectTypeGeneralWaterUse + ' schedule'), @tol)
+    assert_in_epsilon(8760 * get_available_hrs_ratio(unavailable_month_hrs), get_annual_equivalent_full_load_hrs(model, Constants::ObjectTypeMechanicalVentilationHouseFan + ' schedule'), @tol)
+  end
+
   def test_stochastic_schedules
     args_hash = {}
     args_hash['hpxml_path'] = File.absolute_path(File.join(sample_files_dir, 'base-schedules-detailed-occupancy-stochastic.xml'))
@@ -577,6 +609,31 @@ class HPXMLtoOpenStudioSchedulesTest < Minitest::Test
     assert_equal(2, unavailable_schedule_rules.size)
 
     _test_day_schedule(schedule, begin_month, begin_day, year, 0, 24)
+    _test_day_schedule(schedule, end_month, end_day, year, 0, end_hour)
+    _test_day_schedule(schedule, end_month, end_day + 1, year, nil, nil)
+
+    # 2 calendar days, partial first and last day
+    begin_month = 12
+    begin_day = 14
+    begin_hour = 20
+    end_month = 12
+    end_day = 15
+    end_hour = 6
+
+    model, hpxml, _hpxml_bldg = _test_measure(args_hash)
+    year = model.getYearDescription.assumedYear
+
+    schedule = model.getScheduleRulesets.find { |schedule| schedule.name.to_s == sch_name }
+    unavailable_periods = _add_unavailable_period(hpxml, 'Power Outage', begin_month, begin_day, begin_hour, end_month, end_day, end_hour) # note the change of end month/day
+
+    schedule_rules = schedule.scheduleRules
+    Schedule.set_unavailable_periods(model, schedule, sch_name, unavailable_periods)
+    unavailable_schedule_rules = schedule.scheduleRules - schedule_rules
+
+    assert_equal(2, unavailable_schedule_rules.size)
+
+    _test_day_schedule(schedule, begin_month, begin_day - 1, year, nil, nil)
+    _test_day_schedule(schedule, begin_month, begin_day, year, begin_hour, 24)
     _test_day_schedule(schedule, end_month, end_day, year, 0, end_hour)
     _test_day_schedule(schedule, end_month, end_day + 1, year, nil, nil)